feat: apply base information form
