+++ apps/web/src/services/administration/components/RoleUpdateFormBaseInformation.vue
@@ -1,0 +1,85 @@
+<script lang="ts" setup>
+import { watch } from 'vue';
+
+import {
+    PPaneLayout, PHeading, PFieldGroup, PTextInput,
+} from '@spaceone/design-system';
+
+import { i18n } from '@/translations';
+
+import { useFormValidator } from '@/common/composables/form-validator';
+
+interface Props {
+    initialData?: string;
+}
+
+const props = withDefaults(defineProps<Props>(), {
+    initialData: undefined,
+});
+
+const emit = defineEmits<{(e: 'update-validation', after: boolean): void,
+    (e: 'update-form', name: string): void,
+}>();
+
+const {
+    forms: {
+        roleName,
+    },
+    setForm,
+    invalidState,
+    invalidTexts,
+} = useFormValidator({
+    roleName: props.initialData ?? '',
+}, {
+    roleName(value?: string) {
+        if (value === undefined) return '';
+        return value.trim().length > 2 ? '' : i18n.t('IAM.ROLE.FORM.VALIDATION_ROLE_NAME');
+    },
+});
+
+/* Watcher */
+watch(() => roleName.value, (value) => {
+    emit('update-form', value);
+    if (!invalidState.roleName) {
+        emit('update-validation', true);
+    }
+});
+</script>
+
+<template>
+    <p-pane-layout class="role-update-page-base-information">
+        <p-heading heading-type="sub"
+                   :title="$t('IAM.ROLE.DETAIL.BASE_INFORMATION')"
+        />
+        <div class="input-wrapper">
+            <p-field-group
+                :label="$t('IAM.ROLE.DETAIL.NAME')"
+                :invalid="invalidState.roleName"
+                :invalid-text="invalidTexts.roleName"
+                required
+            >
+                <template #default="{invalid}">
+                    <p-text-input class="role-name-input input"
+                                  :value="roleName"
+                                  :invalid="invalid"
+                                  @update:value="setForm('roleName', $event)"
+                    />
+                </template>
+            </p-field-group>
+        </div>
+    </p-pane-layout>
+</template>
+
+<style lang="postcss" scoped>
+.role-update-page-base-information {
+    @apply mx-0;
+    max-width: 100%;
+    .input-wrapper {
+        @apply mx-4 mb-8 flex flex-wrap gap-1 flex-col;
+        .input {
+            max-width: 43.5rem;
+            width: 100%;
+        }
+    }
+}
+</style>
